CP-1104: Move settings controllers
