+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/routing.yml
@@ -29,6 +29,9 @@ akeneo_connectivity_connection_rest:
     prefix: /rest/connections
     trailing_slash_on_root: false
 
+akeneo_connectivity_connection_settings_rest:
+    resource: ./Settings/routing.yml
+
 akeneo_connectivity_connection_webhook_rest:
     resource: ./Webhook/routing.yml
